rename: api 패키지를 interfaces 패키지 하위로 변경
