--- src/main/java/kr/hhplus/be/server/api/point/dto/response/PointResponse.java
@@ -1,17 +1,0 @@
-package kr.hhplus.be.server.api.point.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class PointResponse {
-
-    private Long userId;
-    private Integer balance;
-
-    @Builder
-    private PointResponse(Long userId, Integer balance) {
-        this.userId = userId;
-        this.balance = balance;
-    }
-}
