--- src/main/java/kr/hhplus/be/server/api/product/dto/response/ProductListResponse.java
@@ -1,17 +1,0 @@
-package kr.hhplus.be.server.api.product.dto.response;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.util.List;
-
-@Getter
-public class ProductListResponse<T> {
-
-    private List<T> products;
-
-    @Builder
-    private ProductListResponse(List<T> products) {
-        this.products = products;
-    }
-}
