+++ src/main/java/kr/hhplus/be/server/interfaces/api/coupon/dto/response/CouponListResponse.java
@@ -1,0 +1,19 @@
+package kr.hhplus.be.server.interfaces.api.coupon.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class CouponListResponse {
+
+    private Long userId;
+    private List<CouponResponse> coupons;
+
+    @Builder
+    private CouponListResponse(Long userId, List<CouponResponse> coupons) {
+        this.userId = userId;
+        this.coupons = coupons;
+    }
+}
