+++ src/main/java/kr/hhplus/be/server/interfaces/api/product/dto/response/ProductListResponse.java
@@ -1,0 +1,17 @@
+package kr.hhplus.be.server.interfaces.api.product.dto.response;
+
+import lombok.Builder;
+import lombok.Getter;
+
+import java.util.List;
+
+@Getter
+public class ProductListResponse<T> {
+
+    private List<T> products;
+
+    @Builder
+    private ProductListResponse(List<T> products) {
+        this.products = products;
+    }
+}
